exploration/Matt-Ruso-Sonification/data2music-with-python-part2/requirements.txt
pandas
matplotlib
audiolazy
midiutil
pygame
Pillow
numpy